// lib/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const TaskStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  scraped: 'scraped',
  for_verification: 'for_verification',
  completed: 'completed',
  central_chat: 'central_chat'
} as const

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus]


export const Sender = {
  freelancer: 'freelancer',
  client: 'client'
} as const

export type Sender = (typeof Sender)[keyof typeof Sender]


export const IdeaType = {
  TEXT: 'TEXT',
  URL: 'URL',
  IMAGE: 'IMAGE'
} as const

export type IdeaType = (typeof IdeaType)[keyof typeof IdeaType]